Move network-msg-processing code out of main to its own file
