+++ src/init.cpp
@@ -23,6 +23,7 @@
 #include "miner.h"
 #include "netbase.h"
 #include "net.h"
+#include "net_processing.h"
 #include "policy/policy.h"
 #include "rpc/server.h"
 #include "rpc/register.h"
