+++ src/qt/walletmodel.cpp
@@ -15,8 +15,10 @@
 #include "base58.h"
 #include "keystore.h"
 #include "main.h"
+#include "net.h" // for g_connman
 #include "sync.h"
 #include "ui_interface.h"
+#include "util.h" // for GetBoolArg
 #include "wallet/wallet.h"
 #include "wallet/walletdb.h" // for BackupWallet
