+++ src/test/DoS_tests.cpp
@@ -6,8 +6,8 @@
 
 #include "chainparams.h"
 #include "keystore.h"
-#include "main.h"
 #include "net.h"
+#include "net_processing.h"
 #include "pow.h"
 #include "script/sign.h"
 #include "serialize.h"
